% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(path)
}
\arguments{
\item{path}{Path to search site.}
}
\value{
Tibble with scraped data for all offers from search site.
}
\description{
Combines get_offer_paths and get_data_from_offer_path to facilitate
loading of many offers from search site.
}
