% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_data_from_offer_path}
\alias{get_data_from_offer_path}
\title{Get data from offer path}
\usage{
get_data_from_offer_path(path)
}
\arguments{
\item{path}{Path to the offer site.}
}
\value{
Tibble with scraped data.
}
\description{
Scrap house estate data from given path on otodom.pl
}
