% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_offer_paths}
\alias{get_offer_paths}
\title{Get offer paths}
\usage{
get_offer_paths(path)
}
\arguments{
\item{path}{Path to the search site.}
}
\value{
Paths to offers from search site.
}
\description{
Fetch paths from otodom.pl search site.
}
